Document EnInsect actor (#1184)
* Initial commit

* Go through the rest of the code

* Fix length of EnAnubice.flameCircles

The length used to be 4, but in a for loop 5 elements were used. There
was unused unknown space right after the array and that is now part of
flameCircles.

* Fix offset comment in EnAnubice header

* Fix length of BossGanon2.unk_234

The length used to be 16, but only 15 elements seem to be used. The for
loops only went to 14, and the indicse gotten from a lookup table are
also never higher than 14.

* Revert some ARRAY_COUNTs back to literals

In these cases, using ARRAY_COUNT can make the code more confusing to
read, since the lengths of these arrays are already assumed in other
places in the code or the index is used in multiple arrays.

* Initial commit

* Rename stateTimer to actionTimer

This name is more consistent with actionFunc.

* Name (Setup)Crawl and (Setup)CrawlToSpawn functions

* Rename (Setup)Sink to (Setup)Dig and add (Setup)Sink and (Setup)Drown

* Add IS_DROPPED macro

This is way shorter than checking type == ... || type = ... every time.

* Rename INSECT_FLAG_6 to INSECT_FLAG_LANDED

* Name INSECT_FLAG_0 and the two remaining functions

* Rename INSECT_FLAG_1 to INSECT_FLAG_ON_GROUND

* Replace 'bug' with 'insect' in comment

* Revert changes made in array-count branch

* Update INSECT_DROPPED name in z_player.c

Seems I forgot to rebuild before committing.

* Use bug instead of insect in comments

* Apply suggested changes

* Fix faulty merge

* Run formatter

* Improve `sInitInsectFlags` formatting

* Revert `INSECT_FLAG_ON_GROUND` to `INSECT_FLAG_1`, purpose unclear (to me at least)

* Revert `INSECT_FLAG_TEMP_AND_ALIVE` to `INSECT_FLAG_0`, purpose unclear (to me at least)

* `INSECT_FLAG_TEMP` -> `INSECT_FLAG_IS_SHORT_LIVED`

* `Sink` -> `WalkOnWater`

* `INSECT_FLAG_LANDED` -> `INSECT_FLAG_DROPPED_HAS_LANDED`

* `EnInsect_FoundNearbySoil` -> `EnInsect_TryFindNearbySoil` and cleanup the function

* Run formatter

* Cleanup `actionTimer`

* Some cleanup

* Add `_TYPE_` in `EnInsectType` enum names

* `crawlSoundTimer` -> `crawlSoundDelay`

* Remove `IS_DROPPED` macro

Co-authored-by: Milo van der Tier <milo@vdtier.nl>